multi-threaded kludges
lighttpd is not written to be thread-safe.

These patches may be used for research,
but SHOULD NOT be used in a production environment!

This patch is part of a kludge to have lighttpd run multiple
threads of independent (struct server) instances.

Original patches tried to scale lighttpd to threads,
though justification for such threading was not documented.

This author questions the approach, but these patches aim to port
forward to contemporary lighttpd source code for testing.

Question: could the lighttpd base remain single-threaded
and libmtcp create additional threads for libmtcp work?

x-ref:
  mTCP: A Highly Scalable User-level TCP Stack for Multicore Systems
  https://github.com/mtcp-stack/mtcp