multi-threaded kludges: __thread
This patch is a quick attempt to sprinkle __thread in the code.

Important: lighttpd *has not* been tested to be thread-safe.

Many lighttpd modules and lighttpd features are likely not thread-safe.

lighttpd serving a static file with mod_staticfile appears to work.

OTOH, lighttpd TLS modules have not been modified with __thread
and are not written to be thread-safe.

x-ref:
  mTCP: A Highly Scalable User-level TCP Stack for Multicore Systems
  https://github.com/mtcp-stack/mtcp